Login

+++ src/main/java/io/github/jass2125/projeto/praticas/dao/GameDao.java
@@ -9,15 +9,11 @@ import io.github.jass2125.projeto.praticas.entidades.Game;
 import java.sql.SQLException;
 import java.util.List;
 
-/**
- *
- * @author Anderson Souza
- */
 public interface GameDao {
 
     public List<Game> listarGames() throws SQLException, ClassNotFoundException;
 
     public void salvar(Game game) throws SQLException, ClassNotFoundException;
 
-    public void delete(int id) throws SQLException, ClassNotFoundException;
+    public void delete(Long id) throws SQLException, ClassNotFoundException;
 }
